bagViewer

+++ api/bag_opt_surfaces.c
@@ -283,7 +283,7 @@ bagError bagCreateOptionalDataset (bagHandle bag_hnd, bagData *data, s32 type)
 		case Node_Group:
             memset (&nullNodeGroup, 0, sizeof (bagOptNodeGroup));
             nullNodeGroup.hyp_strength   = BAG_NULL_GENERIC;
-            nullNodeGroup.num_hypotheses = BAG_NULL_GENERIC;
+			nullNodeGroup.num_hypotheses = (u32) BAG_NULL_GENERIC;
 
             status = H5Pset_fill_time  (plist_id, H5D_FILL_TIME_ALLOC);
             status = H5Pset_fill_value (plist_id, datatype_id, &nullNodeGroup);
@@ -304,7 +304,7 @@ bagError bagCreateOptionalDataset (bagHandle bag_hnd, bagData *data, s32 type)
             memset (&nullElevationSolutionGroup, 0, sizeof (bagOptElevationSolutionGroup));
             nullElevationSolutionGroup.shoal_elevation   = BAG_NULL_ELEVATION;
             nullElevationSolutionGroup.stddev        = BAG_NULL_GENERIC;
-            nullElevationSolutionGroup.num_soundings = BAG_NULL_GENERIC;
+            nullElevationSolutionGroup.num_soundings = (u32) BAG_NULL_GENERIC;
 
             status = H5Pset_fill_time  (plist_id, H5D_FILL_TIME_ALLOC);
             status = H5Pset_fill_value (plist_id, datatype_id, &nullElevationSolutionGroup);
